c05 change

+++ ex06/ft_is_prime.c
@@ -6,18 +6,18 @@
 /*   By: yusyamas <yuppiy2000@icloud.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 21:04:43 by yusyamas          #+#    #+#             */
-/*   Updated: 2022/10/18 21:42:08 by yusyamas         ###   ########.fr       */
+/*   Updated: 2022/10/19 10:13:43 by yusyamas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_is_prime(int nb)
 {
-	int	i;
+	long	i;
 
 	if (nb <= 1)
 		return (0);
 	i = 2;
-	while (i * i <= nb)
+	while (i * i <= (long)nb)
 	{
 		if (nb % i == 0)
 			return (0);
